feat: add MyOrders component in page and add metadata
